WIP: Added WCD9378 (mobile mode) support.
Headset playback is working with HPH sequencer override and valid-config bypass settings enabled for now.

AMIC support is currently under debugging. And code cleanup are still in progress.

+++ sound/soc/codecs/wcd-mbhc-v2.h
@@ -223,7 +223,7 @@ struct wcd_mbhc_intr {
 };
 
 struct wcd_mbhc_field {
-	u16 reg;
+	u32 reg;
 	u8 mask;
 };
